fix app bug
