+++ packages/app/components/permissions/PresetItem/IndividualPermissions.tsx
@@ -1,12 +1,10 @@
 import { PermissionCoerced } from "zodiac-roles-sdk"
-import cn from "classnames"
 import { groupPermissions } from "../groupPermissions"
 import TargetItem from "../TargetItem"
-import classes from "./style.module.css"
 import { ChainId } from "@/app/chains"
 import { PermissionsDiff } from "../types"
 import Flex from "@/ui/Flex"
-import AnchorExpandableBox from "./AnchorExpandableBox"
+import IndividualPermissionsExpandable from "./IndividualPermissionsExpandable"
 
 const IndividualPermissions: React.FC<{
   uri: string
@@ -17,16 +15,7 @@ const IndividualPermissions: React.FC<{
   const permissionGroups = groupPermissions(permissions)
 
   return (
-    <AnchorExpandableBox
-      namespace={uri}
-      borderless
-      p={3}
-      className={cn(classes.permissions, BOX_CLASS)}
-      labelCollapsed="Individual permissions"
-      labelExpanded="Individual permissions"
-      toggleClassName={TOGGLE_CLASS}
-      onToggle={diff ? handleToggle : undefined}
-    >
+    <IndividualPermissionsExpandable namespace={uri} inDiffMode={!!diff}>
       <Flex gap={3} direction="column">
         {permissionGroups.map(([targetAddress, permissions]) => (
           <TargetItem
@@ -38,55 +27,8 @@ const IndividualPermissions: React.FC<{
           />
         ))}
       </Flex>
-    </AnchorExpandableBox>
+    </IndividualPermissionsExpandable>
   )
 }
 
 export default IndividualPermissions
-
-const BOX_CLASS = "permissionBox"
-const TOGGLE_CLASS = "permissionBoxToggle"
-export const DIFF_CONTAINER_CLASS = "diffContainer"
-
-// TODO: make this work!
-const handleToggle = (ev: React.MouseEvent<HTMLDivElement, MouseEvent>) => {
-  const diffContainers = [
-    ...document.querySelectorAll(`.${DIFF_CONTAINER_CLASS}`),
-  ]
-  if (diffContainers.length !== 2) {
-    throw new Error("Expected exactly two diff containers")
-  }
-
-  // find clicked permission box
-  const clickedBox = ev.currentTarget.closest(`.${BOX_CLASS}`)
-  if (!clickedBox) {
-    throw new Error("Expected parent permission box")
-  }
-
-  // determine parent diff container
-  const parent = ev.currentTarget.closest(`.${DIFF_CONTAINER_CLASS}`)
-  if (!parent) {
-    throw new Error("Expected parent diff container")
-  }
-
-  // determine index of clicked permission box in parent diff container
-  const clickedIndex = [...parent.querySelectorAll(`.${BOX_CLASS}`)].indexOf(
-    clickedBox
-  )
-
-  // click permission box with same index in other diff container
-  const otherParent = diffContainers.find((c) => c !== parent)
-  const counterpartBox = otherParent?.querySelectorAll(`.${BOX_CLASS}`)[
-    clickedIndex
-  ]
-  if (!counterpartBox) {
-    throw new Error("Expected to find counterpart permission box")
-  }
-  const counterpartToggle = counterpartBox.querySelector(
-    `.${TOGGLE_CLASS}`
-  ) as HTMLElement | null
-  if (!counterpartToggle) {
-    throw new Error("Expected to find counterpart toggle")
-  }
-  counterpartToggle.click()
-}
